+++ client/src/features/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
-import { useAuth } from '../../../stores/AuthProvider';
+import { useAuth } from '../stores/AuthProvider';
 
 interface ProtectedRouteProps {
   children: ReactNode;
