+++ client/src/features/auth/sign-up/PasswordValidation.tsx
