+++ client/src/features/auth/sign-up/SelectRegistrationType.tsx
