+++ client/src/features/auth/sign-up/register-customer.tsx
@@ -4,10 +4,8 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Link } from 'react-router-dom';
-import {
-  PasswordRequirements,
-  passwordValidationSchema,
-} from '../../../components/PasswordValidation';
+
+import {PasswordRequirements, passwordValidationSchema} from '@/features/auth/sign-up/PasswordValidation';
 
 // Zod schema for customer registration validation
 const customerRegistrationSchema = z.object({
