+++ client/src/features/auth/sign-up/register-vendor.tsx
@@ -1,22 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { z } from 'zod';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
 import { Link } from 'react-router-dom';
 import PasswordRequirements, {
   passwordValidationSchema,
-} from '@/app/components/PasswordValidation';
+} from '@/features/auth/sign-up/PasswordValidation';
 
-// Zod schema for shipper registration validation
-const shipperRegistrationSchema = z.object({
+// Zod schema for vendor registration validation
+const vendorRegistrationSchema = z.object({
   username: z
     .string()
     .min(3, 'Username must be at least 3 characters long')
@@ -30,62 +23,47 @@ const shipperRegistrationSchema = z.object({
     .email('Please enter a valid email address')
     .min(1, 'Email is required'),
   password: passwordValidationSchema,
-  assignedHub: z.string(),
+  businessName: z
+    .string()
+    .min(2, 'Business name must be at least 2 characters long')
+    .max(100, 'Business name must be less than 100 characters'),
+  businessAddress: z
+    .string()
+    .min(5, 'Business address must be at least 5 characters long')
+    .max(200, 'Business address must be less than 200 characters'),
   profilePicture: z.string().optional(),
 });
 
-interface ShipperRegistrationData {
-  username: string;
-  email: string;
-  password: string;
-  assignedHub?: string; // Allow undefined for initial state
-  profilePicture?: string;
-}
+type VendorRegistrationData = z.infer<typeof vendorRegistrationSchema>;
 
-export interface RegisteredShipper {
+export interface RegisteredVendor {
   id: string;
   username: string;
-  email: string;
-  assignedHub?: string;
+  businessName: string;
+  businessAddress: string;
   profilePicture: string;
   role: string;
 }
 
-interface RegisterShipperProps {
-  onRegistrationSuccess?: (Shipper: RegisteredShipper) => void;
+interface RegisterVendorProps {
+  onRegistrationSuccess?: (vendor: RegisteredVendor) => void;
   onSwitchToLogin?: () => void;
 }
 
-export const RegisterShipper: React.FC<RegisterShipperProps> = ({
+export const RegisterVendor: React.FC<RegisterVendorProps> = ({
   onRegistrationSuccess,
 }) => {
-  const [formData, setFormData] = useState<ShipperRegistrationData>({
+  const [formData, setFormData] = useState<VendorRegistrationData>({
     username: '',
     email: '',
     password: '',
-    assignedHub: undefined, // Start with undefined to show placeholder
+    businessName: '',
+    businessAddress: '',
     profilePicture: '',
   });
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string>('');
   const [fieldErrors, setFieldErrors] = useState<Record<string, string>>({});
-  const [hubs, setHubs] = useState<
-    { _id: string; hubName: string; hubLocation: string }[]
-  >([]);
-
-  useEffect(() => {
-    fetch('http://localhost:5001/api/hubs')
-      .then((res) => res.json())
-      .then((data) => {
-        // Sort hubs alphabetically by hubName for consistent ordering
-        const sortedHubs = data.hubs.sort(
-          (a: { hubName: string }, b: { hubName: string }) =>
-            a.hubName.localeCompare(b.hubName),
-        );
-        setHubs(sortedHubs);
-      })
-      .catch((err) => console.error('Failed to load hubs:', err));
-  }, []);
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target;
@@ -103,39 +81,13 @@ export const RegisterShipper: React.FC<RegisterShipperProps> = ({
     }
   };
 
-  const handleSelectChange = (value: string) => {
-    setFormData((prev) => ({
-      ...prev,
-      assignedHub: value,
-    }));
-    // Clear general error and field-specific error when user selects
-    if (error) setError('');
-    if (fieldErrors.assignedHub) {
-      setFieldErrors((prev) => ({
-        ...prev,
-        assignedHub: '',
-      }));
-    }
-  };
-
   const validateForm = (): boolean => {
     try {
       // Clear previous field errors
       setFieldErrors({});
 
-      // Check for undefined assignedHub first
-      if (!formData.assignedHub) {
-        setFieldErrors({ assignedHub: 'Please select an assigned hub' });
-        setError('Please select an assigned hub');
-        return false;
-      }
-
-      // Validate using Zod schema with complete data
-      const completeData = {
-        ...formData,
-        assignedHub: formData.assignedHub,
-      };
-      shipperRegistrationSchema.parse(completeData);
+      // Validate using Zod schema
+      vendorRegistrationSchema.parse(formData);
       return true;
     } catch (error) {
       if (error instanceof z.ZodError) {
@@ -174,12 +126,13 @@ export const RegisterShipper: React.FC<RegisterShipperProps> = ({
         username: formData.username,
         email: formData.email,
         password: formData.password,
-        assignedHubId: formData.assignedHub!, // Send as assignedHubId to match backend expectation
+        businessName: formData.businessName,
+        businessAddress: formData.businessAddress,
         profilePicture: formData.profilePicture,
       };
 
       const response = await fetch(
-        'http://localhost:5001/api/auth/register/shipper',
+        'http://localhost:5001/api/auth/register/vendor',
         {
           method: 'POST',
           headers: {
@@ -192,14 +145,14 @@ export const RegisterShipper: React.FC<RegisterShipperProps> = ({
       const data = await response.json();
 
       if (response.ok) {
-        console.log('Shipper registration successful:', data);
+        console.log('Vendor registration successful:', data);
 
-        // Store Shipper data and token in localStorage
-        localStorage.setItem('Shipper', JSON.stringify(data.Shipper));
+        // Store vendor data and token in localStorage
+        localStorage.setItem('vendor', JSON.stringify(data.vendor));
         localStorage.setItem('token', data.token);
 
-        onRegistrationSuccess?.(data.Shipper);
-        alert('Shipper registration successful!');
+        onRegistrationSuccess?.(data.vendor);
+        alert('Vendor registration successful!');
       } else {
         setError(data.message || 'Registration failed');
         console.error('Registration failed:', data);
@@ -232,9 +185,7 @@ export const RegisterShipper: React.FC<RegisterShipperProps> = ({
           <form onSubmit={handleSubmit} className="space-y-6">
             {/* Error Message */}
             {error && (
-              <div className="bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded-lg text-sm">
-                {error}
-              </div>
+              <div className=" px-4 py-3 rounded-lg text-sm">{error}</div>
             )}
 
             {/* Username Field */}
@@ -275,32 +226,43 @@ export const RegisterShipper: React.FC<RegisterShipperProps> = ({
               )}
             </div>
 
-            {/* Assigned Distribution Hub Field */}
+            {/* Business Name Field */}
             <div className="space-y-2">
-              <Label htmlFor="assignedHub">Assigned distribution hub *</Label>
-              <Select
-                value={formData.assignedHub}
-                onValueChange={handleSelectChange}
+              <Label htmlFor="businessName">Business Name *</Label>
+              <Input
+                type="text"
+                id="businessName"
+                name="businessName"
+                value={formData.businessName}
+                onChange={handleInputChange}
+                placeholder="Enter your business name"
                 disabled={loading}
+                className={`w-full ${fieldErrors.businessName ? 'border-red-500' : ''}`}
                 required
-              >
-                <SelectTrigger
-                  className={`w-full ${fieldErrors.assignedHub ? 'border-red-500' : ''}`}
-                >
-                  <SelectValue placeholder="Select your assigned hub" />
-                </SelectTrigger>
-                <SelectContent>
-                  {hubs.map((hub) => (
-                    <SelectItem key={hub._id} value={hub._id}>
-                      {hub.hubName} - {hub.hubLocation}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-              {fieldErrors.assignedHub && (
+              />
+              {fieldErrors.businessName && (
                 <p className="text-red-500 text-sm">
-                  {fieldErrors.assignedHub}
+                  {fieldErrors.businessName}
                 </p>
+              )}
+            </div>
+
+            {/* Business Address Field */}
+            <div className="space-y-2">
+              <Label htmlFor="businessAddress">Business Address *</Label>
+              <Input
+                type="text"
+                id="businessAddress"
+                name="businessAddress"
+                value={formData.businessAddress}
+                onChange={handleInputChange}
+                placeholder="Enter your business address"
+                disabled={loading}
+                className={`w-full ${fieldErrors.businessAddress ? 'border-red-500' : ''}`}
+                required
+              />
+              {fieldErrors.businessAddress && (
+                <p className=" text-sm">{fieldErrors.businessAddress}</p>
               )}
             </div>
 
@@ -319,7 +281,7 @@ export const RegisterShipper: React.FC<RegisterShipperProps> = ({
                 required
               />
               {fieldErrors.password && (
-                <p className="text-red-500 text-sm">{fieldErrors.password}</p>
+                <p className=" text-sm">{fieldErrors.password}</p>
               )}
 
               {/* Password Requirements */}
@@ -332,7 +294,7 @@ export const RegisterShipper: React.FC<RegisterShipperProps> = ({
               disabled={loading}
               className="w-full py-3 rounded-lg font-medium"
             >
-              {loading ? 'Creating Account...' : 'Create Shipper Account'}
+              {loading ? 'Creating Account...' : 'Create Vendor Account'}
             </Button>
 
             {/* Sign In Link */}
